Fix compilation on 64-bit systems.
I'm not crazy about this use of pointers as keys, but it seems to be
okay for the time being. One or two don't make sense, though, so I've
made a comment about that in the code.

+++ gtex/patch_id_texture.C
@@ -29,7 +29,9 @@
 void 
 PatchIDStripCB::faceCB(CBvert* v, CBface* f)
 {
-   uint rgba = IDRefImage::key_to_rgba(uint(f->patch()));
+   /* FIXME: QuLogic: I don't understand how this will work, since f->patch()
+    * is a pointer, and the key is some internal thing. */
+   uint rgba = IDRefImage::key_to_rgba(uintptr_t(f->patch()));
    glColor4ubv((GLubyte*)&rgba);
 
    glVertex3dv(v->loc().data());
